Added audio conversion and ability to play simultanious sounds.
Now need to separate these into their own classes and clean everything
up. Make a class for the converter if the file is unplayable, then i
need to return the playable clip and run that in a loop in a new thread
for the required amount of time, or use the system shown here to run
them in the same thread cocurrently, need to configure timing however.